+++ lib/ts/framework/awsLambda/framework.ts
@@ -214,7 +214,10 @@ export class AWSResponse extends BaseResponse {
         }
     };
 
-    sendResponse = (response: APIGatewayProxyResult | APIGatewayProxyStructuredResultV2) => {
+    sendResponse = (response?: APIGatewayProxyResult | APIGatewayProxyStructuredResultV2) => {
+        if (response === undefined) {
+            response = {};
+        }
         let headers:
             | {
                   [header: string]: boolean | number | string;
@@ -303,17 +306,25 @@ export const middleware = (handler?: Handler): Handler => {
         try {
             let result = await supertokens.middleware(request, response);
             if (result) {
-                return response.sendResponse({});
+                return response.sendResponse();
             }
             if (handler !== undefined) {
                 let handlerResult = await handler(event, context, callback);
                 return response.sendResponse(handlerResult);
             }
-            return response.sendResponse({});
+            /**
+             * it reaches this point only if the API route was not exposed by
+             * the SDK and user didn't provide a handler
+             */
+            response.setStatusCode(404);
+            response.sendJSONResponse({
+                error: `The middleware couldn't serve the API path ${request.getOriginalURL()}, method: ${request.getMethod()}. If this is an unexpected behaviour, please create an issue here: https://github.com/supertokens/supertokens-node/issues`,
+            });
+            return response.sendResponse();
         } catch (err) {
             await supertokens.errorHandler(err, request, response);
             if (response.responseSet) {
-                return response.sendResponse({});
+                return response.sendResponse();
             }
             throw err;
         }
@@ -321,7 +332,7 @@ export const middleware = (handler?: Handler): Handler => {
 };
 
 export interface AWSFramework extends Framework {
-    middleware: () => Handler;
+    middleware: (handler?: Handler) => Handler;
 }
 
 export const AWSWrapper: AWSFramework = {